\name{compare_meth}
\alias{compare_meth}
\title{
compare methylation between smoothed and raw methylation data
}
\description{
compare methylation between smoothed and raw methylation data
}
\usage{
compare_meth(cr, chr, start, end, x = NULL, x2 = NULL)
}
\arguments{

  \item{cr}{cr}
  \item{chr}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{x}{cr with smoothed methylation}
  \item{x2}{cr with raw methylation}

}
\examples{
# There is no example
NULL
}
