\name{cr_hilbert}
\alias{cr_hilbert}
\title{
Hilbert curve visualization of CRs
}
\description{
Hilbert curve visualization of CRs
}
\usage{
cr_hilbert(cr, template, txdb, chromosome = paste0("chr", 1:22), merge_chr = TRUE)
}
\arguments{

  \item{cr}{cr}
  \item{template}{template}
  \item{txdb}{txdb}
  \item{chromosome}{chromosome}
  \item{merge_chr}{whether merge chromsomes into one plot}

}
\examples{
# There is no example
NULL
}
