\name{cr_scatterplot_me}
\alias{cr_scatterplot_me}
\title{
scatter plot between methylation and expression
}
\description{
scatter plot between methylation and expression
}
\usage{
cr_scatterplot_me(cr, expr, gi = NULL, text_column = NULL,
    xlab = "Methylation", ylab = "Expression")
}
\arguments{

  \item{cr}{correlated regions from \code{\link{correlated_regions}}}
  \item{expr}{expression}
  \item{gi}{gene id}
  \item{text_column}{-text_column}
  \item{xlab}{xlab}
  \item{ylab}{ylab}

}
\examples{
# There is no example
NULL
}
