\name{enriched_heatmap_list_on_genomic_features}
\alias{enriched_heatmap_list_on_genomic_features}
\title{
Enriched heatmap on a certain gf
}
\description{
Enriched heatmap on a certain gf
}
\usage{
enriched_heatmap_list_on_genomic_features(cr, gf, hm_list = NULL, hm_name = NULL, ...)
}
\arguments{

  \item{cr}{cr}
  \item{gf}{genomic features}
  \item{hm_list}{hm_list}
  \item{hm_name}{hm_name}
  \item{...}{pass to}

}
\examples{
# There is no example
NULL
}
