\name{enriched_heatmap_list_on_tss_cgi}
\alias{enriched_heatmap_list_on_tss_cgi}
\title{
Enriched heatmap on TSS-CGI
}
\description{
Enriched heatmap on TSS-CGI
}
\usage{
enriched_heatmap_list_on_tss_cgi(cr, cgi, txdb, expr, hm_list = NULL, hm_name = NULL, by = "gene", ...)
}
\arguments{

  \item{cr}{cr}
  \item{cgi}{CpG Islands}
  \item{txdb}{txdb}
  \item{expr}{expr}
  \item{hm_list}{hm_list}
  \item{hm_name}{hm_name}
  \item{by}{by}
  \item{...}{pass to}

}
\examples{
# There is no example
NULL
}
