\name{plot_expr_distribution}
\alias{plot_expr_distribution}
\title{
distribution of expression
}
\description{
distribution of expression
}
\usage{
plot_expr_distribution(expr, log = "none", label = NULL, main = NULL)
}
\arguments{

  \item{expr}{expressed matrix}
  \item{log}{log2 or log10}
  \item{label}{type of the expression}
  \item{main}{title}

}
\details{
it generates two figures, only make plot in [q0, q90]
}
\examples{
# There is no example
NULL
}
