\name{reduce2}
\alias{reduce2}
\title{
reduce based on the width of the interval
}
\description{
reduce based on the width of the interval
}
\usage{
reduce2(gr, max_gap = 1000, gap = bp(1000))
}
\arguments{

  \item{gr}{genomic regions}
  \item{max_gap}{max gap}
  \item{gap}{if it is a value less than 1, it is the ratio of the interval}

}
\examples{
# There is no example
NULL

}
