\name{reduce_cr}
\alias{reduce_cr}
\title{
refuce cr regions
}
\description{
refuce cr regions
}
\usage{
reduce_cr(cr, expr, txdb, max_gap = 1000, gap = 1.0, mc.cores = 1)
}
\arguments{

  \item{cr}{cr }
  \item{expe}{expression}
  \item{txdb}{txdb}
  \item{max_gap}{maximum gap}
  \item{gap}{gap}
  \item{mc.cores}{number of cores}

}
\details{
pos_CR and neg_CR are reduced separatedly
}
\examples{
# There is no example
NULL
}
