\name{systemdf}
\alias{systemdf}
\title{
Wrapper of system calls in which input and output are all table-like files
}
\description{
Wrapper of system calls in which input and output are all table-like files
}
\usage{
systemdf(cmd, envir = parent.frame(), verbose = FALSE)
}
\arguments{

  \item{cmd}{shell command}
  \item{envir}{environment where to look for variables encoded in \code{cmd}}
  \item{verbose}{whether print messages}

}
\details{
This function (system + data frame) provides a convinient way to invoke
system calls in R. Since most of system calls expect tables as input and output, 
\code{\link{systemdf}} does following things step by step:

\itemize{
  \item use backtick to mark variables which are data frames or other variables which can be converted to data frames by \code{\link[base]{as.data.frame}}
  \item extract data frames
  \item write data frames into temporary files 
  \item replace variables names with paths that correspond to temporary files
  \item make the system call 
  \item finally send back the output by piping back to R
}

A simple example is as follows:

  \preformatted{
    bed1 = circlize::generateRandomBed(nr = 1000)
    bed2 = circlize::generateRandomBed(nr = 1000)
    df = systemdf("bedtools closest -a `bed1` -b `bed2` | awk '$1==\"chr1\"'")  }
}
\value{
A data frame. If there is no line generated by the system call, the returned value is a three-column
data frame with 0 row.
}
\examples{
# There is no example
NULL

}
