\name{cr_overlap_to_genomic_features}
\alias{cr_overlap_to_genomic_features}
\title{
enrichment between cr and genomic feature
}
\description{
enrichment between cr and genomic feature
}
\usage{
cr_overlap_to_genomic_features(cr, gf_list, species = NULL, chromosome = paste0("chr", 1:22))
}
\arguments{

  \item{cr}{cr}
  \item{gf_list}{gf_list}
  \item{species}{species}
  \item{chromosome}{chromosomes}

}
\examples{
# There is no example
NULL

}
