\name{extract_field_from_gencode}
\alias{extract_field_from_gencode}
\title{
extract field from gencode GTF file
}
\description{
extract field from gencode GTF file
}
\usage{
extract_field_from_gencode(file, level = "gene", primary_key = "gene_id", field = "gene_name")
}
\arguments{

  \item{file}{path of the gencode GTF file}
  \item{level}{level }
  \item{primary_key}{primary key}
  \item{field}{field}

}
\examples{
# There is no example
NULL
}
